pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vnstock-mcp"
version = "0.1.0"
description = "Vnstock Comprehensive API MCP Server"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastmcp",
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["vnstock_mcp"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
